iOS build should not use a large Gigacage capacity regardless of MACH_VM_MAX_ADDRESS. https://bugs.webkit.org/show_bug.cgi?id=298542 rdar://160131122
Reviewed by Keith Miller.

The range of the virtual addresses does not necessarily imply that the VA space is available to
be allocated from.

No new test needed because this is only enforcing current behavior.  We're just making this
behavior (in terms of Gigacage size selection) explicit.

* Source/bmalloc/bmalloc/Gigacage.h:

Canonical link: https://commits.webkit.org/299729@main