Remove work around for DisplayOperatorMinHeight https://bugs.webkit.org/show_bug.cgi?id=297646 rdar://159260284
Reviewed by Frédéric Wang.

This patch aligns WebKit with Gecko / Firefox and Web Specification [1]:

This patch removes workaround when DisplayOperatorMinHeight is too small
(e.g., in case of Cambria Math font on Windows), for non-Apple ports, we
have updated harfbuzz with fix. So this workaround was not necessary anymore.

It progresses web platform test, which shows that we are not compensating
for the constant being too small.

[1] https://w3c.github.io/mathml-core/#layout-of-operators

* Source/WebCore/rendering/mathml/MathOperator.cpp:
(WebCore::MathOperator::calculateDisplayStyleLargeOperator):
* LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/display-operator-min-height-no-workaround-expected.txt: Progress

Canonical link: https://commits.webkit.org/299773@main