[VideoMediaSampleRenderer] Take into account the amount on in-flight compressed samples when determining if we have reached enough data. https://bugs.webkit.org/show_bug.cgi?id=298639 rdar://160256931
Reviewed by Youenn Fablet.

The calculation of how many compressed samples were pending was racy,
as soon as we decoded the first frame, it would be reset to the actual number
of frames in the compressed queue, which at the time is always 1.
The next read would then not take into account any of the in-flight waiting
for the tasks to run.

As a consequence, we would enqueue a high number of compressed frames.

* Source/WebCore/platform/graphics/cocoa/VideoMediaSampleRenderer.h:
* Source/WebCore/platform/graphics/cocoa/VideoMediaSampleRenderer.mm:
(WebCore::VideoMediaSampleRenderer::areSamplesQueuesReadyForMoreMediaData const):
(WebCore::VideoMediaSampleRenderer::compressedSamplesCount const):
(WebCore::VideoMediaSampleRenderer::enqueueSample):
(WebCore::VideoMediaSampleRenderer::decodeNextSampleIfNeeded):
(WebCore::VideoMediaSampleRenderer::flushCompressedSampleQueue):

Canonical link: https://commits.webkit.org/299798@main