Quirks are not always correctly computed, for instance for reloaded documents https://bugs.webkit.org/show_bug.cgi?id=298596 rdar://160197523
Reviewed by Brent Fulgham.

Quirks may be determined early in the Document cycle, for instance when parsing is finished and quirks are checked for styling.
At that time, the Document may not be attached to a Frame, which may lead to the Document having an empty top document URL.
This will prevent correct computation of some quirks, which will remain for the lifetime of the document.

To fix this, we recompute quirks when a document becomes a top document.
Manually tested.

* Source/WebCore/dom/Document.cpp:
(WebCore::Document::didBecomeCurrentDocumentInFrame):
* Source/WebCore/page/Quirks.h:

Canonical link: https://commits.webkit.org/299804@main