REGRESSION (iOS 26): editing/selection/ios/selection-handles-in-readonly-input.html is a constant timeout https://bugs.webkit.org/show_bug.cgi?id=298679 rdar://156727121
Reviewed by Abrar Rahman Protyasha.

Changes in Gestures.framework in iOS 26 broke the tap-and-half gesture to
perform selections in WebKit. rdar://160314418 tracks a fix at the system level.

The fact that the tap-and-half gesture no longer selects text causes
editing/selection/ios/selection-handles-in-readonly-input.html to timeout.
To restore test coverage, rewrite the test to select text by tapping and
dragging.

Then, a new test is added to provide coverage specifically for the tap-and-half
gesture. Once rdar://160314418 is fixed, the skip expectation can be removed.

* LayoutTests/editing/selection/ios/select-text-using-tap-and-half-expected.txt: Added.
* LayoutTests/editing/selection/ios/select-text-using-tap-and-half.html: Added.
* LayoutTests/editing/selection/ios/selection-handles-in-readonly-input.html:
* LayoutTests/platform/ios/TestExpectations:

Canonical link: https://commits.webkit.org/299828@main