[WPE][cross-toolchain-helper] Unable to build with USE_SYSTEM_SYSPROF_CAPTURE=ON after 299709@main https://bugs.webkit.org/show_bug.cgi?id=298713
Unreviewed build-fix.

299709@main updated the sysprof version on the yocto env and switched
USE_SYSTEM_SYSPROF_CAPTURE back to enabled by default.

However this caused a build falure, because sysprof-capture-4 is a static
library (.a), and yocto by default ships that kind of libraries on a -staticdev
package which is not included by default along with the headers on the sysroot.

Fix that by explicitly including the sysprof-staticdev package in this case,
and also backport a fix for issues when unloading the capture library.

* Tools/yocto/meta-openembedded_and_meta-webkit.patch:

Canonical link: https://commits.webkit.org/299856@main