Avoid unnecessary ref churn in identifierForStyleProperty in EditingStyle.cpp https://bugs.webkit.org/show_bug.cgi?id=298705
Reviewed by Vitor Roriz and Chris Dumez.

The static analyzer was getting confused by `RefPtr fontStyleValue`. Because it comes from `auto`,
it couldn't deduce that we're using `RefPtr` as the actual template so it gives up determining the type.
This PR fixes eliminates this ambiguity by explicitly declaring the type of `value` and removes the
superfluous construction of Ref in calling fontStyleAngleFromCSSFontStyleWithAngleValueDeprecated.

* Source/WebCore/editing/EditingStyle.cpp:
(WebCore::identifierForStyleProperty):

Canonical link: https://commits.webkit.org/299860@main