AX: Trigger timeout in 'waitFor' function if promise could not be resolved https://bugs.webkit.org/show_bug.cgi?id=298590
Reviewed by Tyler Wilcock.

When evaluating an expression in 'expectAsync' or 'expectAsyncExpression', if
the expression evaluated is false or triggers an exception, the operator 'await'
will hang waiting for the promise to be resolved or rejected. As a consequence
of that, many subtests that may fail are actually timing out.

To prevent 'waitFor' to hang forever, a timeout timer was added. The timer will
resolve the promise if it could not be resolved within a time threshold
(3 seconds).

* LayoutTests/platform/glib/TestExpectations:
* LayoutTests/resources/accessibility-helper.js:

Canonical link: https://commits.webkit.org/299867@main