[GTK][WPE][OpenXR] Check GBM device before attempting to use it https://bugs.webkit.org/show_bug.cgi?id=298721
Reviewed by Adrian Perez de Castro.

I don't expect this to happen in a properly configured environment, but
if GBMDevice::create fails (because the can't open the DRM nodes),
m_gbmDevice will be null, and attempting to dereference it in
OpenXRCoordinator::createGLDisplay leads to a crash. This adds a null
check to avoid that.

* Source/WebKit/UIProcess/XR/openxr/PlatformXROpenXR.cpp:
(WebKit::OpenXRCoordinator::createGLDisplay const):

Canonical link: https://commits.webkit.org/299886@main