Broaden Swift memory safety to all of WebKit https://bugs.webkit.org/show_bug.cgi?id=297589 rdar://158673557
Reviewed by Mike Wyrzykowski and Elliott Williams

This moves the Strict Memory Safety Swift compiler flag from WebKit.xcconfig to
an alternative config file that applies to nearly all WebKit targets.

It also broadens the set of flags we use to match those that WebGPU has
determined to be best.

This is the second attempt to land this change. The first one ran into an error
in older versions of the Swift @Test macro where it emitted unsafe code
(rdar://151238560). As it happens, WebKit's continuous integration
infrastructure already has a fix for that in place, but that fix isn't
being picked up due to some other problem where Swift is sometimes
using an older version of the macro dylib (rdar://159858481). This needs
to be resolved before we land this commit again. But, those problems
are unrelated to the content of this commit so it's identical to what
we landed last time.

Canonical link: https://commits.webkit.org/299897@main