[JSC] Do not save / restore SIMD registers when Wasm SIMD is disabled https://bugs.webkit.org/show_bug.cgi?id=298778 rdar://160471660
Reviewed by Dan Hecht.

When wasm SIMD is not enabled, there is an assertion which ensures we
are not using SIMD registers for scratch register allocator code. This
assertion is too restrictive, but for now, just bypass it by not storing
SIMD registers when wasm SIMD is disabled.

* Source/JavaScriptCore/wasm/WasmThunks.cpp:
(JSC::Wasm::materializeBaselineDataGenerator):

Canonical link: https://commits.webkit.org/299901@main