[CG] radius of CGGaussianBlurStyle is not in the BaseCTM coordinates https://bugs.webkit.org/show_bug.cgi?id=298725 rdar://160383014
Reviewed by Simon Fraser.

In /299740@main, the radius of the CGGaussianBlurStyle was scaled by the device
scale factor as a work around for a CG bug on accelerated drawing.

The radius should not be scaled by the device scale factor. In other words it
should be in the BaseCTM coordinates. This is the behavior of the unaccelerated
drawing.

The workaround should stay for the accelerated drawing till the CG bug is fixed.

* LayoutTests/css3/filters/effect-unaccelerated-graphics-context-blur-expected.html: Added.
* LayoutTests/css3/filters/effect-unaccelerated-graphics-context-blur.html: Added.
* Source/WebCore/platform/graphics/filters/FEGaussianBlur.cpp:
(WebCore::FEGaussianBlur::createGraphicsStyle const):

Canonical link: https://commits.webkit.org/299903@main