[AutoFill Debugging] Make `WKTextExtractionLinkItem`'s URL nullable https://bugs.webkit.org/show_bug.cgi?id=298822 rdar://160535084
Reviewed by Richard Robinson.

Make the `url` property on this object nullable, and exclude it from the text extraction if it's
`nil`. From code inspection, it should be possible for a non-empty `WTF::URL` to be passed into this
codepath which yields a `nil` `NSURL`. I discovered a single crash report underneath this codepath
while testing, but could not reproduce it on subsequent attempts.

* Source/WebKit/UIProcess/API/Cocoa/_WKTextExtraction.swift:
(WKTextExtractionLinkItem.url):
(WKTextExtractionLinkItem.textRepresentationParts):
* Source/WebKit/UIProcess/API/Cocoa/_WKTextExtractionInternal.h:

Canonical link: https://commits.webkit.org/299932@main