[GTK][WPE] WebXR tests fail to be run in EWS due to assertions https://bugs.webkit.org/show_bug.cgi?id=298726
Reviewed by Dan Glastonbury.

Some of the assertions in WebXROpaqueFramebuffer do not make sense when testing the WebXR API
because we're using a fake XR device that does not render anything, so it does not provide
platform images/surfaces for WebGL to render into.

That was preventing us from enabling WebXR testing as some EWS run with all ASSERTs enabled
(even with release builds). We can pass a isTesting parameter via the structs used to
exchange XR information between the UI and Web processes that will be used later in
conditional assertions. That parametter is not used for production code so we don't
need to encode/decode it in IPC.

This allows us to run almost 100 WebXR passing tests and to remove old expectations that
are no longer needed as those tests run as expected.

* LayoutTests/imported/w3c/web-platform-tests/webxr/ar-module/idlharness.https.window-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/webxr/hand-input/idlharness.https.window-expected.txt: Added.
* LayoutTests/platform/glib/TestExpectations:
* LayoutTests/platform/wpe/http/wpt/webxr/xrDevice_requestSession_previously_enabled_feature_not_requested.https-expected.txt: Removed.
* LayoutTests/platform/wpe/imported/w3c/web-platform-tests/webxr/events_referenceSpace_reset_immersive.https-expected.txt: Removed.
* LayoutTests/platform/wpe/imported/w3c/web-platform-tests/webxr/events_referenceSpace_reset_inline.https-expected.txt: Removed.
* LayoutTests/platform/wpe/imported/w3c/web-platform-tests/webxr/xrDevice_disconnect_ends.https-expected.txt: Removed.
* LayoutTests/platform/wpe/imported/w3c/web-platform-tests/webxr/xrSession_input_events_end.https-expected.txt: Removed.
* Source/WebCore/Modules/webxr/WebXROpaqueFramebuffer.cpp:
(WebCore::WebXROpaqueFramebuffer::startFrame):
(WebCore::WebXROpaqueFramebuffer::blitShared):
(WebCore::WebXROpaqueFramebuffer::blitSharedToLayered):
(WebCore::WebXROpaqueFramebuffer::setupFramebuffer):
* Source/WebCore/Modules/webxr/WebXROpaqueFramebuffer.h:
* Source/WebCore/platform/xr/PlatformXR.h:
* Source/WebCore/testing/WebFakeXRDevice.cpp:
(WebCore::SimulatedXRDevice::frameTimerFired):

Canonical link: https://commits.webkit.org/299948@main