REGRESSION(299031@main): `RELEASE_ASSERT(a.globalPosition() != b.globalPosition())` is hit under `compareAnimationsByCompositeOrder()` when calling `WebAnimation::commitStyles()` https://bugs.webkit.org/show_bug.cgi?id=298860 rdar://160101397
Reviewed by Anne van Kesteren.

In 299031@main we changed our implementation for `Animation.commitStyles()` to no longer go through the animation target's associated
keyframe effect stack and instead use all animations associated with that target in order to include animations that would be marked
as not relevant [0] since the spec behavior changed to include animations that would be right at their finished time.

That change meant that we would potentially iterate over idle animations. In itself, this is not an issue since idle animations would
not have an effect on the animated style. However, it caused an assertion in `compareAnimationsByCompositeOrder()` to fail because in
some cases we would try to sort style-originated animations that did not yet have a global position and would have been canceled by the
time `commitStyles()` was called, as was the case in the newly-created test.

We now filter the list of animations to consider to remove any idle animation.

[0] https://drafts.csswg.org/web-animations-1/#relevant-animations-section

Test: webanimations/commit-style-crash.html
* LayoutTests/webanimations/commit-style-crash-expected.txt: Added.
* LayoutTests/webanimations/commit-style-crash.html: Added.
* Source/WebCore/animation/WebAnimation.cpp:
(WebCore::WebAnimation::commitStyles):

Canonical link: https://commits.webkit.org/299952@main