Improve upon IOSurfacePool's DUMP_POOL_STATISTICS https://bugs.webkit.org/show_bug.cgi?id=298808 rdar://160504843
Reviewed by Simon Fraser.

The GPU process will create at least one IOSurfacePool per WebContent process it serves.
When debugging the pool in the presence of multiple tabs the output of
DUMP_POOL_STATISTICS becomes ambiguous. This change adds an object identifier
for each IOSurfacePool so the log output becomes possible to follow in this scenario.

No tests as this is not a behavior change in normal builds.

* Source/WebCore/platform/graphics/cg/IOSurfacePool.cpp:
    Add pool identifiers to all DUMP_POOL_STATISTICS logs. Also
    computes the percentage of total surfaces which are purgeable
    and should not count against the process in the vm ledger.
    To help denote pool creation/destruction I also added some logging
    to the ctor/dtor.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/graphics/cg/IOSurfacePool.h:
* Source/WebCore/platform/graphics/cg/IOSurfacePoolIdentifier.h: Added.
    This is simply an object identifier we can use to disambiguate pool statistics
    when there is more than one pool in the GPUP.

Canonical link: https://commits.webkit.org/299971@main