Assign interactionID to simulated clicks rdar://159873276 https://bugs.webkit.org/show_bug.cgi?id=298409
Reviewed by Ryosuke Niwa.

When computing interactionID for click events, we now consider whether
the click is being simulated by a keypress. This behavior is required
for some WPT tests and is consistent with Chrome and Firefox, but is
currently not in accordance with the spec. An issue on the spec has
been filed about this: https://github.com/w3c/event-timing/issues/147.

To implement this, a data member containing the interaction ID was
added to UIEvent.

This patch also adds a workaround to a bug with IME keydowns on
finalizeEventTimingEntry(). See https://bugs.webkit.org/show_bug.cgi?id=165004

Now passing:
* LayoutTests/imported/w3c/web-platform-tests/event-timing/interactionid-keyboard-event-simulated-click-button-space-expected.txt
* LayoutTests/imported/w3c/web-platform-tests/event-timing/interactionid-keyboard-event-simulated-click-checkbox-space-expected.txt
* LayoutTests/imported/w3c/web-platform-tests/event-timing/interactionid-keyboard-event-simulated-click-link-enter-expected.txt

Canonical link: https://commits.webkit.org/299973@main