Script 'update-angle' assumess target Webkit commit is always main https://bugs.webkit.org/show_bug.cgi?id=298928
Reviewed by Kimmo Kinnunen.

At it is, the script doesn't work in a scenario where we're trying to
bisect a past ANGLE interval update.  The script assumes the target
WebKit commit is always 'main'.  When bisecting a past ANGLE update
interval we'd like the script doesn't see ANGLE changes before a
certain WebKit commit.

To achieve that, the target WebKit commit is set to HEAD, which works
both if the current commit is tip of main or a past commit.

In addition, the shebang was changed to '/usr/bin/env bash' since strictly
talking this is a Bash script (the script uses double brackets and read -r,
only available in Bash).  A spared line was also removed.

* Tools/Scripts/update-angle:

Canonical link: https://commits.webkit.org/300020@main