Broaden Swift memory safety to all of WebKit https://bugs.webkit.org/show_bug.cgi?id=297589 rdar://158673557
Unreviewed, relanding following internal build fix.

This moves the Strict Memory Safety Swift compiler flag from WebKit.xcconfig to
an alternative config file that applies to nearly all WebKit targets.

It also broadens the set of flags we use to match those that WebGPU has
determined to be best.

This is the third attempt to land this change. The first one ran into an error
in older versions of the Swift @Test macro where it emitted unsafe code
(rdar://151238560). WebKit suffered from this problem due to parts of our CI
using an older version of the macro dylib (rdar://159858481).

We attempted to reland once we believed all SDKs were fixed, but we
were wrong. We are now more confident that we no longer have versions
of toolchains suffering from this bug so this is a second reland.

This reverts commit 92f0c1d77416da479c5cacce164e56a6466edee0,
which was the revert of the original fix plus a follow-up code change.

Canonical link: https://commits.webkit.org/300022@main