[JSC] WASM IPInt SIMD: x86_64: implement bitwise instructions https://bugs.webkit.org/show_bug.cgi?id=298956 rdar://160694776
Reviewed by Yusuke Suzuki.

Implement x86_64 IPInt bitwise instructions.

For bitmask, which is already emulated, rearrange register assignments
so that t3, which is %rcx on x86_64, is the shift-by. That avoids
offlineasm having to insert xchg before/after the shift to get more
efficient code. Also, there is no need to move sp to another register.

This allows simd-instructions-bitwise.js test to pass on x86_64.

Testing: On Intel and ARM:
    jsc --validateOptions=1 --useDollarVM=true  --useConcurrentJIT=false
        --useWasmIPIntSIMD=1 -m simd-instructions-bitwise.js

Canonical link: https://commits.webkit.org/300043@main